# backend/docs/swagger/complete-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vertex Markets API",
    "version": "1.0.0",
    "description": "Production-ready backend API for a modern Forex/Trading company website. Provides authentication, user management, CMS content management, markets, trading platforms, blog, contact management, newsletter subscriptions, careers/jobs, partner management, file uploads, admin dashboard, and global search capabilities.",
    "contact": {
      "name": "Vertex Markets Development Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.vertexmarkets.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/v1/login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "JWT access token stored in HTTP-only cookie"
      }
    },
    "schemas": {
...
      "AuthRegisterRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "minLength": 8, "example": "securePassword123" },
          "profile": {
            "type": "object",
            "properties": {
              "firstName": { "type": "string", "example": "John" },
              "lastName": { "type": "string", "example": "Doe" },
              "phone": { "type": "string", "example": "+1234567890" }
            }
          },
          "role": { "type": "string", "enum": ["Admin", "Customer"], "example": "Customer" }
        }
      },
      "AuthRegisterResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "Registered successfully" },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "example": "user_abc123" },
                  "email": { "type": "string", "example": "user@example.com" }
                }
              },
              "verificationToken": { "type": "string", "example": "ev_abc123..." }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user account",
        "description": "Create a new user account with email and password. Optionally accepts profile information and role. Returns a verification token for email verification.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthRegisterRequest" },
              "example": {
                "email": "user@example.com",
                "password": "securePassword123",
                "profile": {
                  "firstName": "John",
                  "lastName": "Doe",
                  "phone": "+1234567890"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthRegisterResponse" }
              }
            }
          },
          "400": { "description": "Validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationError" } } } },
          "409": { "description": "Email already in use", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Conflict" } } } }
        }
      }
    },
...
    "/api/v1/partners/register": {
      "post": {
        "tags": ["Partners"],
        "summary": "Register as a partner",
        "description": "Submit a partner registration request. Public endpoint.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PartnerRegisterRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Partner registration submitted" },
          "400": { "description": "Validation failed" },
          "409": { "description": "Email already registered" }
        }
      }
    },
    "/api/v1/partners/referral": {
      "post": {
        "tags": ["Partners"],
        "summary": "Submit a client referral",
        "description": "Submit a client referral through a partner. Public endpoint.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReferralSubmitRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Referral submitted" },
          "400": { "description": "Validation failed" }
        }
      }
    },
    "/api/v1/partners/status/{id}": {
      "get": {
        "tags": ["Partners"],
        "summary": "Check partner registration status",
        "description": "Public endpoint to check the status of a partner registration.",
        "security": [],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Partner status" },
          "404": { "description": "Partner not found" }
        }
      }
    },
    "/api/v1/partners": {
      "get": {
        "tags": ["Partners Admin"],
        "summary": "List all partners (admin)",
        "description": "Retrieve a paginated list of all partner registrations.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "minimum": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "in": "query", "name": "sortBy", "schema": { "type": "string" } },
          { "in": "query", "name": "sortOrder", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "in": "query", "name": "q", "schema": { "type": "string" } },
          { "in": "query", "name": "status", "schema": { "type": "string", "enum": ["pending", "under_review", "approved", "rejected"] } },
          { "in": "query", "name": "partnerType", "schema": { "type": "string", "enum": ["Introducing Broker (IB)", "Affiliate Partner", "Regional Partner", "Business Partner", "Institutional Partner"] } },
          { "in": "query", "name": "assignedTo", "schema": { "type": "string" } },
          { "in": "query", "name": "active", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "List of partners" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/v1/partners/{id}": {
      "get": {
        "tags": ["Partners Admin"],
        "summary": "Get partner by ID (admin)",
        "description": "Retrieve a specific partner registration by ID.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Partner details" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Partner not found" }
        }
      },
      "put": {
        "tags": ["Partners Admin"],
        "summary": "Replace partner (admin)",
        "description": "Fully replace a partner registration.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": { "type": "string" },
                  "companyName": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string" },
                  "country": { "type": "string" },
                  "city": { "type": "string" },
                  "website": { "type": "string", "format": "url" },
                  "partnerType": { "type": "string", "enum": ["Introducing Broker (IB)", "Affiliate Partner", "Regional Partner", "Business Partner", "Institutional Partner"] },
                  "experience": { "type": "string" },
                  "monthlyClients": { "type": "string" },
                  "businessDescription": { "type": "string" },
                  "referralSource": { "type": "string" },
                  "status": { "type": "string", "enum": ["pending", "under_review", "approved", "rejected"] },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Partner updated" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Partner not found" }
        }
      },
      "patch": {
        "tags": ["Partners Admin"],
        "summary": "Partially update partner (admin)",
        "description": "Partially update a partner registration.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": { "type": "string" },
                  "companyName": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string" },
                  "country": { "type": "string" },
                  "city": { "type": "string" },
                  "website": { "type": "string", "format": "url" },
                  "partnerType": { "type": "string", "enum": ["Introducing Broker (IB)", "Affiliate Partner", "Regional Partner", "Business Partner", "Institutional Partner"] },
                  "experience": { "type": "string" },
                  "monthlyClients": { "type": "string" },
                  "businessDescription": { "type": "string" },
                  "referralSource": { "type": "string" },
                  "status": { "type": "string", "enum": ["pending", "under_review", "approved", "rejected"] },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Partner patched" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Partner not found" }
        }
      },
      "delete": {
        "tags": ["Partners Admin"],
        "summary": "Soft-delete partner (admin)",
        "description": "Soft-delete a partner registration.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Partner deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Partner not found" }
        }
      }
    },
    "/api/v1/partners/{id}/status": {
      "patch": {
        "tags": ["Partners Admin"],
        "summary": "Update partner status (admin)",
        "description": "Update the approval status of a partner registration.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["pending", "under_review", "approved", "rejected"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status updated" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Partner not found" }
        }
      }
    },
    "/api/v1/partners/{id}/assign": {
      "patch": {
        "tags": ["Partners Admin"],
        "summary": "Assign partner (admin)",
        "description": "Assign a partner registration to an admin for processing.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assignedTo"],
                "properties": {
                  "assignedTo": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Partner assigned" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Partner not found" }
        }
      }
    },
    "/api/v1/partners/{id}/notes": {
      "patch": {
        "tags": ["Partners Admin"],
        "summary": "Update partner notes (admin)",
        "description": "Update internal notes on a partner registration.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["notes"],
                "properties": {
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Notes updated" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Partner not found" }
        }
      }
    },
    "/api/v1/partners/referrals": {
      "get": {
        "tags": ["Partners Admin"],
        "summary": "List all referrals (admin)",
        "description": "Retrieve a paginated list of all client referrals submitted by partners.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "minimum": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "in": "query", "name": "sortBy", "schema": { "type": "string" } },
          { "in": "query", "name": "sortOrder", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "in": "query", "name": "q", "schema": { "type": "string" } },
          { "in": "query", "name": "status", "schema": { "type": "string", "enum": ["new", "contacted", "qualified", "converted", "closed"] } },
          { "in": "query", "name": "referredBy", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of referrals" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/v1/partners/referrals/{id}": {
      "get": {
        "tags": ["Partners Admin"],
        "summary": "Get referral by ID (admin)",
        "description": "Retrieve a specific referral by its ID.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Referral details" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Referral not found" }
        }
      },
      "delete": {
        "tags": ["Partners Admin"],
        "summary": "Soft-delete referral (admin)",
        "description": "Soft-delete a client referral.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Referral deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Referral not found" }
        }
      }
    },
...
  },
...
}